src: move code inside src/cleanpython
update imports in tests
add __init__.py
Issue #59
